requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3>=1.34.0,<2
python-multipart>=0.0.9
openai>=1.40.0,<2
fastembed>=0.5.0,<1
pypdf>=4.0.0,<6
